import from top-up-debugger repo without repackaging
This branch is a backward-compatibility branch. Master was also imported
from the same place, but this branch differs from master in that it does
not re-package for building with Maker.  All OOPSLA16 artifacts use the
older Maker, and expect libedb to have been built without Maker. Hence
they will use this branch of libedb.